fixes compilation errors in Graph from changes to isConnectionValid function and change of isOccupiedThreshold parameter to double type

+++ lab3/src/Graph.h
@@ -22,7 +22,7 @@ public:
 
 	void generate_connections(int connectionsPerNode, double maxDistance);
 
-	void prune_invalid_connections(nav_msgs::OccupancyGrid map, double robotSize, int isOccupiedThreshold);
+	void prune_invalid_connections(nav_msgs::OccupancyGrid map, double robotSize, double isOccupiedThreshold);
 
 private:
 	double calculate_distance(Node start, Node end);
@@ -31,7 +31,7 @@ private:
 
 	int convertPositionToGridIndex(double position, double mapLowerLimit, double resolution);
 
-	bool isConnectionValid(Node startNode, Node endNode, nav_msgs::OccupancyGrid map, double robotSize, int isOccupiedThreshold);
+	bool isConnectionValid(Node startNode, Node endNode, nav_msgs::OccupancyGrid map, double robotSize, double isOccupiedThreshold);
 
 	void bresenham(int x0, int y0, int x1, int y1, std::vector<int>& x, std::vector<int>& y);
 };
